pyproject.toml
[project]
name = "portals-client"
version = "0.1.0"
description = ""
authors = [
    {name = "Niron Koren",email = "dev@tammuz.co"}
]
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "web3 (>=7.12.0,<8.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "eth-account (>=0.13.7,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
